IDF v5.0 compatibility
- Changes to datatypes in logging function calls for compiling on IDF v5.0
- mbedTLS API usage updated to work with v2 and v3 of mbedTLS
- OTA third-party dependency, `cbor`, is now added using IDF Component Manager

+++ examples/jobs/main/mqtt_demo_helpers.c
@@ -945,8 +945,8 @@ BaseType_t xPublishToTopic( MQTTContext_t * pxMqttContext,
         }
         else
         {
-            LogInfo( ( "PUBLISH sent for topic %.*s to broker with packet ID %u.\n\n",
-                       topicFilterLength,
+            LogInfo( ( "PUBLISH sent for topic %.*s to broker with packet ID %d.\n\n",
+                       (int) topicFilterLength,
                        pcTopicFilter,
                        outgoingPublishPackets[ ucPublishIndex ].packetId ) );
